# docs/swagger.yaml
definitions:
...
  model.Pagination:
    properties:
      page:
        type: integer
      size:
        type: integer
      totalItem:
        type: integer
      totalPage:
        type: integer
    type: object
...
  model.UserLogin:
    properties:
      email:
        maxLength: 255
        type: string
      password:
        maxLength: 255
        minLength: 8
        type: string
      tokenCaptcha:
        minLength: 100
        type: string
    required:
    - email
    - password
    - tokenCaptcha
    type: object
  model.UserResponse:
    properties:
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      phoneNumber:
        type: string
      profilePicture:
        type: string
      role:
        type: string
    type: object
  model.UserUpdatePassword:
    properties:
      confirmPassword:
        type: string
      oldPassword:
        maxLength: 255
        minLength: 8
        type: string
      password:
        maxLength: 255
        minLength: 8
        type: string
    required:
    - confirmPassword
    - oldPassword
    - password
    type: object
  utility.PaginationResponse:
    properties:
      data: {}
      pagination: {}
    type: object
  utility.ResponseError:
    properties:
      error:
        type: string
    type: object
  utility.ResponseSuccess:
    properties:
      data: {}
    type: object
info:
  contact: {}
paths:
...
  /api/user:
    get:
      description: Search for users by name, role, phone number, and email
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Name
        in: query
        name: name
        type: string
      - description: Phone number
        in: query
        name: phoneNumber
        type: string
      - description: Email
        in: query
        name: email
        type: string
      - description: Role
        in: query
        name: role
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utility.PaginationResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.UserResponse'
                  type: array
                pagination:
                  items:
                    $ref: '#/definitions/model.Pagination'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utility.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utility.ResponseError'
      summary: Search for users
      tags:
      - User
    post:
      consumes:
      - multipart/form-data
      description: Create a new user with name, phone number, email, and role
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: User name
        in: formData
        name: name
        required: true
        type: string
      - description: Phone number
        in: formData
        name: phoneNumber
        required: true
        type: string
      - description: Email
        in: formData
        name: email
        required: true
        type: string
      - description: Profile picture
        in: formData
        name: profilePicture
        type: file
      - description: Role (admin, journalist)
        in: formData
        name: role
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/utility.ResponseSuccess'
            - properties:
                data:
                  $ref: '#/definitions/model.UserResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utility.ResponseError'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/utility.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utility.ResponseError'
      summary: Create a new user
      tags:
      - User
  /api/user/{id}:
    delete:
      description: Delete an existing user by their ID
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utility.ResponseSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utility.ResponseError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utility.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utility.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utility.ResponseError'
      summary: Delete user by ID
      tags:
      - User
    get:
      description: Retrieve a specific user by their ID
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utility.ResponseSuccess'
            - properties:
                data:
                  $ref: '#/definitions/model.UserResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utility.ResponseError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utility.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utility.ResponseError'
      summary: Get user by ID
      tags:
      - User
    put:
      consumes:
      - multipart/form-data
      description: Update an existing user's details
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: User name
        in: formData
        name: name
        required: true
        type: string
      - description: Phone number
        in: formData
        name: phoneNumber
        required: true
        type: string
      - description: Email
        in: formData
        name: email
        required: true
        type: string
      - description: Profile picture
        in: formData
        name: profilePicture
        type: file
      - description: Password
        in: formData
        name: password
        type: string
      - description: Role
        in: formData
        name: role
        required: true
        type: string
      - description: Delete profile picture
        in: formData
        name: deleteProfilePicture
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utility.ResponseSuccess'
            - properties:
                data:
                  $ref: '#/definitions/model.UserResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utility.ResponseError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utility.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utility.ResponseError'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/utility.ResponseError'
      summary: Update user by ID
      tags:
      - User
  /api/user/current:
    get:
      description: Get the profile of the currently logged-in user
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utility.ResponseSuccess'
            - properties:
                data:
                  $ref: '#/definitions/model.UserResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utility.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utility.ResponseError'
      summary: Get current user's profile
      tags:
      - User
  /api/user/current/password:
    patch:
      consumes:
      - application/json
      description: Update the current user's password
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: New password details
        in: body
        name: password
        required: true
        schema:
          $ref: '#/definitions/model.UserUpdatePassword'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/utility.ResponseSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utility.ResponseError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utility.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utility.ResponseError'
      summary: Update current user's password
      tags:
      - User
  /api/user/current/profile:
    patch:
      consumes:
      - multipart/form-data
      description: Update the current user's name, phone number, email, and profile
        picture
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: User name
        in: formData
        name: name
        required: true
        type: string
      - description: Phone number
        in: formData
        name: phoneNumber
        required: true
        type: string
      - description: Email
        in: formData
        name: email
        required: true
        type: string
      - description: Profile picture
        in: formData
        name: profilePicture
        type: file
      - description: Delete profile picture
        in: formData
        name: deleteProfilePicture
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utility.ResponseSuccess'
            - properties:
                data:
                  $ref: '#/definitions/model.UserResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utility.ResponseError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utility.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utility.ResponseError'
      summary: Update current user's profile
      tags:
      - User
  /api/user/login:
    post:
      consumes:
      - application/json
      description: User logs in with email/phone number and password, returning a
        JWT token
      parameters:
      - description: User login data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.UserLogin'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utility.ResponseSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utility.ResponseError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utility.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utility.ResponseError'
      summary: User login
      tags:
      - User
swagger: "2.0"